Reuse docker container creation for build
